+++ client/src/components/course-results-view/CourseResultsTable.js
@@ -18,34 +18,7 @@ import CourseResultsTableToolbar from './CourseResultTableToolbar';
 import CourseResultsTableHead from './CourseResultsTableHead';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
-
-const descendingComparator = (a, b, orderBy) => {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-};
-
-const getComparator = (order, orderBy) => {
-  return order === 'desc'
-    ? (a, b) => descendingComparator(a, b, orderBy)
-    : (a, b) => -descendingComparator(a, b, orderBy);
-};
-
-const stableSort = (array, comparator) => {
-  const stabilizedThis = array.map((el, index) => [el, index]);
-  stabilizedThis.sort((a, b) => {
-    const order = comparator(a[0], b[0]);
-    if (order !== 0) {
-      return order;
-    }
-    return a[1] - b[1];
-  });
-  return stabilizedThis.map((el) => el[0]);
-};
+import sortingServices from '../../services/sorting';
 
 const CourseResultsTable = ({ attainments, students, calculateFinalGrades, updateGrades }) => {
 
@@ -107,8 +80,8 @@ const CourseResultsTable = ({ attainments, students, calculateFinalGrades, updat
               attainments = {attainments}
             />
             <TableBody>
-              {stableSort(studentsToShow,
-                getComparator(order, orderBy))
+              { sortingServices.stableSort(studentsToShow,
+                sortingServices.getComparator(order, orderBy))
                 .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
                 .map((student) => {
